Update README, hero section, and page theme
Decided to change the theme of the webpage from a Apple store theme to a video game store theme. Any reference to Apple using images or text was changed as a result. This change is because I like video games more than I like Apple products, and it allows me to be more creative and flexible in setting up the page with what I want.

THe hero section of the webpage is pretty much done, and I'm thinking of moving on to the other sections.

I forget to add this to commit d0ad9c6, but the README was intially changed there too.